Implement summary_model.py
Working on the Summary Model. Still issues with the interactive pie object - specifically with scrolling that has an issue with the underlying scrolling module in the kivy library.

Also have changed the database connection to different schema. Instead using campus_movements_v2. This is a simpler version than original version.